requirements.txt
# Core dependencies
google-generativeai>=0.3.0
pandas>=1.5.0
PyPDF2>=3.0.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=22.0.0
flake8>=5.0.0